Patches by Russell King:
Provide a new function to encapsulate the actions that are needed
when GPS inforamtion is updated.  Move GpsUpdated out of the
NMEAParser class - XCSoar.cpp is the only file which really needs
this.

Note: devCAI302.cpp was only partly triggering a vario update, this
change fixes that oversight.

JMW added update trigger in simulator when dragging the screen, to improve
snappiness.