Patch by Russell King
We don't need two sets of almost identical code to drive two ports,
one set is sufficient.  We can do this with one set of code and two
data instances.

Comparing Port.cpp with Port2.cpp, it seems that there's a number of
minor differences between the two - Port2.cpp contains an additional
call to SetupComm().  Maybe a bug fix missing from Port.cpp?

C standard for preprocessing directives requires the '#' at the
start of the line.  There can then be spaces after the '#' before
the directive.  Clean this up as well.