We don't need two sets of almost identical code to drive two ports, one set is sufficient.  We can do this with one set of code and two data instances.
Comparing Port.cpp with Port2.cpp, it seems that there's a number of
minor differences between the two - Port2.cpp contains an additional
call to SetupComm().  Maybe a bug fix missing from Port.cpp?

C standard for preprocessing directives requires the '#' at the
start of the line.  There can then be spaces after the '#' before
the directive.  Clean this up as well.

----

As DeviceDescriptor_t contains a pointer to the ComPort class, we
can invoke its Close method directly from devClose().  ComPorts
will be shut down automatically when we close the device.

-----
Move the COM port initialisation code into devInit().  This allows
device.cpp to become the interface between XCSoar, the COM ports
and the device drivers.  This is to prepare for the next two
steps...

-----

Add locking using the commport critical section, to ensure that
 d->Com can not unexpectedly change beneath any other threads.

Add a note about this locking - there is the potential for deadlock
here.  That being the possibility of an AB:BA deadlock between the
comport lock and the flight data lock.

When we want to stop the com port rx thread, we hold have LockComm()'d.
The com port rx thread may be waiting on obtaining the flight data
lock, which we may be holding - resulting in a warning about the rx
thread not being stopped.

Solve this by ensuring that, whenever there's the possibility of us
taking the comm port lock inside the flight data lock, we first
take the comm port lock outside.  Since Critical sections are
recursive for the same thread, this avoids the possibility of deadlock.