Patches by Russell King
[PATCH] Separate out common logger declaration code

No point duplicating the logger code for each device.  Move it
into a separate function, and call it with the device we want
to send the declaration to.

[PATCH] Clean up logger declaration code.

Having three entry points for logger declarations into the device
drivers is messy and rather inflexible.  Instead, create one
method to handle a declaration, and call it with a structure
containing all the information it needs to create a declaration
for the attached device.

This also means we can sensibly use LockComm() in devices.cpp
for this, rather than in Logger.cpp.

[PATCH] Move fSimMode duplication to devices.cpp

All of the drivers avoid trying to communicate with attached devices
when they're called in simulator mode.  It's pointless having this
logic duplicated in all the drivers; move it up one level to
device.cpp