experimental masked window class
functions but I don't like it yet.  ideally the canvas itself should
be able to create a mask canvas for temporary use, but we don't want every
canvas to do so.  will try out a method whereby a canvas can have a mask
on demand and deleted on exit.  the motivation behind all this is
to simplify the otherwise complicated DrawBitmapIn and and_or_scale type method
which is obtuse.

what we want is canvas.draw_centered_masked_bitmap(x,y,width,height);