MapWindow, PaintWindow: set the IDC_ARROW cursor
When PaintWindow started to register its own class instead of reusing
"STATIC", the mouse cursor began to disappear sometimes.  This patch
explicitly sets WNDCLASS.hCursor to IDC_ARROW.  We could do it in
MainWindow, too, but that isn't worth it because all of MainWindow is
tiled with child windows anyway.