experimental privacy of mutexes via client class
access to the mutex is only via being a friend or
a subclass of the client class.  This could be a useful
way to ensure one class doesn't have access to a lock
it shouldn't.  Have to be careful though because exported
functions that are public that invoke the locks grant privelige
to the outside world.