MapWindow: resize Window on full screen toggle
When not in full screen mode, don't let MapWindow span the whole
screen, drawing the info boxes on top of it.  Limit the size of
MapWindow to the actual map area on the screen.  This will allow us to
disable most of the artificial clipping, and let GDI decide on its own
where to clip.