RWLock: fall back to Mutex on Pocket PC
The Poco WIN32 implementation of the RWLock doesn't work on Windows
Mobile, because that OS doesn't support
WaitForMultipleObjects(fWaitAll=true):

 http://msdn.microsoft.com/en-us/library/aa931008.aspx

 "fWaitAll [in] Unsupported; set to FALSE."

This patch is a hotfix: fix the assertion failure by switching to a
pure Mutex backend.  Note that the "try" functions aren't implemented
properly.  In the long run, we need a better solution, but for now,
this is enough to repair the Windows Mobile build.