fix under-3-char Asset number bug. Caused OLC to reject flight.
The issue was that if for any reason XCSoar thinks the PDA's asset ID
is not 3 chars, it was using 1 or 2 chars, then appending the date
to create the first line of the IGC file.  But, because the array had
only two chars, it added a '\0' as char 3 which caused the string append
to fail, putting the date on the second line of the IGC file, confusing
the OLC by the incorrect format.