Fix NAVWarning bug so never log bad points to IGC File
The IGC spec indicates that B records should never be
logged if the logger is returning a gpsinvalid status.
XCSoar stores this status in the gps_info.NAVWarning flag.
I moved the check for this in the LogPointToFile
which is the final routine that writes B record fixes
to the IGC file so this rule will never be violated.
Note that for other log records (e.g. the FRecord) we DO
want to record as needed during this state.