Remove TMP.IGC and post-flight move file error handling logic
Previousely, XCSoar wrote the IGC file to the root of the device
in a file called "/TMP.IGC" and when the logger was stopped, it
copied the file to the XCSoarData folder.  The drawback was that
if the PDA crashed, there was not log file.  Worse, if the pilot
restarted the logger, it would immediately overright the existing
TMP.IGC file making it impossible to recover, merge the two IGC
files etc.
In 5.2.4 we added error handling so it detected that a prior version
of the TMP.IGC file existed (e.g. the PDA had crashed), and warned
the user, moved files around.  This was ugly.
Now, we're writing directly to the final .IGC file in the XCSoarData
folder so if the PDA crashes, it will create a new file with a
new name and the pilot can append the two files after the flight
just like he would if a IGC-certified logger crashed.
Note that there is another patch that adds buffering so that
we're not writing to the SD card every time we log a point.