Add IGC write buffering.
This is non-stdio because stdio buffering on PDAs adds
junk to end of file if simulated battery failure.
We believe this is a bug in the OS of PDAs because
the "junk" is only observed on the PDAs, not on desktop
or wine versions of the program.
The purpose of this is to that we don't write directly
to an SD card every time we log a point which causes
excessive wear on the SDCard physical structure.
Also note that there is  another patch
that corrects the DiskBufferFlush routnine in this patch
so the DiskBufferFlush DOES uses stdio to write the 10
records all in a single physical write.