LoggerSign.cpp, LoggerImpl.hpp: use stdio in IGC buff flush
Alter igc buffering to use stdio buffer during flush.
The purpose of the IGC Buffering is to reduce
the number of physical writes to SD cards which
have lifespan limited by the number of writes.  We build a
custom buffer for this (see note about PDA OS bug that
appends junk to IGC file if PDA crash while using buffered
IO and leaving file open).  In our custom "Flush" routine,
we open the IGC file and write ~10 records all at once and
close it.  Because we write w/ stdio, it is buffered
and only creates about 1-2 physical writes.