Registry: added DeviceConfig::port_type
SERIAL (for serial ports) is the only option available currently, and
is the default.  This patch includes a change for dlgConfiguration to
prepend a list of port types in the port field.  We could make that a
separate field in the configuration dialog, but that would probably be
more confusing for our users.