AutoForceFinalGlide: removed this parameter
The parameter is confusing and shouldn't be required.  Current
behaviour is as if this value is set to true.

It's much simpler to have the display switch to final glide mode
whenever the aircraft is above final glide, irregardless of whether
there are many or just one turnpoint remaining.

Simpler behaviour is more easily understood.