NMEA/Info: convert base classes to members
Prepare for the "big cleanup": all attributes should be grouped in
sub-structs, each of them with well-defined semantics.  Big problem
was that without searching, you could never know from which base
struct an attribute was inherited.

Some of this will eventually be undone; "Time" will probably return to
NMEA_INFO, and many of the AIRCRAFT_STATE attributes will be moved
somewhere else.  The design of NMEA_INFO should not be limited by the
requirements of the task engine.