MenuBar: return focus after mouse capture is released
A ButtonWindow (as implemented by Windows) grabs the keyboard focus
when the user starts dragging it, but doesn't return the focus after
the user has stopped dragging.  Focus on menu buttons isn't useful, we
should rather return it to the map window.  This patch returns it to
the button's parent window (MainWindow), which in turn forwards it to
the MapWindow.  This is the second part of the fix for ticket #224.