LogFile: Removed logfile lock
Since the application is quitting without saying goodbye as soon as Log*() is called with a line break, the locking doesn't seem to be a good idea here. Additionally there is no need for a lock since, the logfile isn't that important and if the logfile is already opened by another thread, the fopen will just fail anyway.