IO: reimplementation of the TextReader code
The generic base class is "Source", which provides a buffered stream
of records (char or TCHAR).  FileSource and ZipSource just provide
byte streams, and lines are splitted inside LineSplitter.  The
conversion to TCHAR is implemented in yet another class, called
ConvertLineReader.  All of these glued together give you the same
features as the old FileTextReader and ZipTextReader classes.