MapWindowEvents.cpp: fix Virtual Keys
Virtual Keys allow up/down events for devices with no
up/down hardware buttons.
They are enabled / disabled by a configuration setting currently in the
last configuration "Experimental" screen but need to be moved to
the "Interface screen."  The "Help Text" on the configuration
screen describes how to use them.