LoggerImpl.cpp, .hpp: correct Invalid fix handling
The correct procedure for Invalid fixes is to repeat the gps coords of the last
valid fix with the current timestamp.
Also added logic to exclude the "SetHome()" location at startup so if the logger is
started prior to the GPS getting a fix, it does not log XCSoar's "home" point in the IGC file.