removed broken classes RasterMapRaw, RasterMapCache
It seems that nobody ever used RasterMapRaw and RasterMapCache, and
both have been broken for several months already - because the
TERRAIN_INFO struct was changed in a way that's not even portable on
two builds of the same target (FIXED_MATH).  It would be too tedious
to fix them, so let's remove them.