TerrainRenderer: fix slope shading and offset bugs
With steep terrain shading, overflow of magnitude of the slope vector caused
erroneous slope to be generated.  This is now fixed, alongwith rationalisation of the
class (removal of unnecessary items).  This fixes TRAC ticket #156.

Also fixed offset error.  Note work still to be performed to reenable faster traversal of
height buffer, since it was found to be buggy.