DrawThread: configure low priority for this thread
The DrawThread is the heaviest thread, and letting it dominate the
weak embedded CPU may lead to a sluggish feel.  It is more important
to make the UI responsive, and since nobody has to wait for the
DrawThread, let's give it a lower priority.

This adds the methods Thread::set_low_priority().