Data/Dialogs: use the default font for all lists
Use MapWindowFont instead of MapWindowBoldFont.  Lists are packed with
information, and a bold font isn't more readable - actually, it's the
opposite.  If we really want a bigger font, we should increase the
font size, not make it bold.  That is probably true for all form
controls, but this patch only cares for WndListFrame.