FileReader: reactivate Find*FlashCard() on Windows CE
When we switched to gcc, the Find*FlashCard() code was disabled by an
#ifdef, which caused a lot of confusion among users, because XCSoar
couldn't find their data files anymore.  This patch cleans up the code
a little, and reactivates the code.  Unfortunately, version 0.59.1 of
mingw32ce's header "projects.h" has a syntax error, a closing curly
brace is missing, which this patch adds.  As soon as mingw32ce
corrects this bug, we must remove it.