GlueMapWindowEvents: Implemented MouseGestures
(see http://en.wikipedia.org/wiki/Pointing_device_gesture)

If you press the mouse button over the map and "draw" a gesture it can activate the desired action (if recognized)

A gesture is defined by a string of up to ten of these four characters:
U (draw a line up), D (draw a line down), L (draw a line left) and R (draw a line right)


A gesture string like "LDR" would indicate that the user painted something that looked close to the letter C.
To draw the gesture "LDR" you need to press the mouse button, move the mouse to the left for some distance, then move the mouse down and afterward to the right again. After that release the mouse button.

This basic implementation includes the following gestures:
U - emulate a press of the hardware UP key
D - emulate a press of the hardware DOWN key
L - emulate a press of the hardware LEFT key
R - emulate a press of the hardware RIGHT key
DU - open the main menu
DR - open the goto dialog