NMEA/Derived: Removed duplicate TeammateCode[Valid] variables
These variables are also included in the SettingsComputer and are usually not calculated but entered by hand. If the FLARM provides the values it may write to the SettingsComputer.

This patch merges/fixes the old teamcode and the flarm teamcode. Before this patch(set) only one of them (flarm) was really working because TeammateCode was only read from one location and not the other one.