Screen/Animation: delete this library
In my opinion, the animations drawn by the DrawWireRects library are
pretty useless, and only waste resources.  During a flight, the pilot
will not enjoy eye candy.

Besides that, we were violating the author's copyright, because the
orignal code was licensed under the "Code Project Open License", and
relicensing to GPL is forbidden.