DeviceBlackboard: Fixed EnergyHeight()
As the formula states the EnergyHeight is calculated by v^2/(2*g).
There is no need to take the mc speed into account in this.
If auto mc is enabled, the auto mc depends on the thermal avg calculations, which depend on energy height, which depends on the mc speed, which depends on auto mc again.
This couldn't work and as a result now the auto mc is fixed and so are the TC Avg, Gain, Avg30 infoboxes.