InfoBoxes: Added InfoBoxContent base class and InfoBoxFactory
The InfoBoxContent class will be the base class of all future ContentProviders.
The Update() function will be called with the InfoBoxWindow as an argument where e.g. SetComment() can be called.
Constructors of specific implementations of the ContentProvider class may save a type variable so that groups of related infoboxes can share the same code.

InfoBoxFactory::Create() is a static factory function that will return the right ContentProvider depending on the given type of InfoBox.