MapDrawHelper: use Canvas::copy_and() to copy the airspace buffer
Attempt to emulate some transparency with the "SRCAND" blit
operation.  This aims to improve the readability of objects "behind"
the airspace shade.  It turned out that by only switching to SRCAND,
the airspace shades get too dark, so the new function light_color()
masks the color to be light enough.

This fixes TRAC ticket #283.