Makefile: generate GNU gettext template
This patch adds rules to generate the GNU gettext template
po/xcsoar.pot.  Although this is a generated file, it is maintained in
git.  That's a hack, but this is how gettext is best used.

We could theoretically import the existing translations (German and
Italian), but I don't think it's worth the effort, because at least
the German translation quality isn't acceptable.