Driver/CAI302: add 180 degrees to wind direction
The CAI302 emits the direction the wind moves to, not the direction
from which the wind originates.  During my flights, I always thought
XCSoar displayed it all wrong, but on a strong wind day, it was
obvious that XCSoar miscalculated the wind - which was because it read
the wrong direction from my CAI302 in the first place.  What a dumb
bug!