TaskStats: move computing code to separate class
Instances of the TaskStats class are open copied by the assignment
operator.  This is a complex operation, because the DistanceStat class
incorporates an AvFilter object.  AvFilter needs to do heap allocation
when copied.  These values are not needed outside of AbstractTask.
This patch moves all of the internal attributes which are only need to
calculate the stats to a separate class.  We have to do this for three
classes: DistanceStat, ElementStat and TaskStats.