Math/fixed: remove explicit assignment operator
The automatic assignment operator generated by the C++ compiler does
the same.  A major disadvantage of the explicit operator is that it is
a barrier for automatic memcpy() in larger structs; without it, the
C++ compiler may use memcpy() to assign larger structs containing
"fixed" objects.

This patch reduces the WM5 binary by 30 kB.