LocalPath: use flash disk if XCSoarData exists
If the user creates XCSoarData on a flash disk, XCSoar shall use this
directory as the primary data path, overriding CSIDL_PERSONAL.  This
is useful when you install XCSoar from the CAB file in main memory,
but all data should be stored on the flash disk.  The second major use
case is auto-run from a SD card - Windows will copy the EXE file to
main memory, and run it from there.