Data/XCSoar.rc,build/resource: Implements new SVG icons.
New icons added to resource file.

SVG graphics are now rendered using the x/y zoom parameters; icon
dimensions are determined by the canvas size of the SVG file.

The ideal display for the baseline 320x240 resolution is assumed to
be a 4" display (100 pixels per inch). Icons rendered at higher
resolutions are given the suffix -x where -x represents the
required display ppi for the icon to have the same physical
dimensions as the ideal 320x240 display. For example, on a 640x480
5" display 160ppi icons would appear the same physical size as
100ppi icons on a 320x240 4" display.