Thermal.cpp: remove deadlock condition
Handlekey routine was locking the TaskManager and then
concurrently locking the Blackboard (in the call to SetMC).
This is the opposite order of lock requests used by the
Calculation thread.  The Calculation thread first locks
the Blackboard, and while it is locked, it then locks the
TaskManager.  There is no need for the HandleKey routine
to maintain a lock on Task Manager for the entire routine.
So we now release the lock before requesting the lock on
the Blackboard.

This deadlock condition can be reproduced in XCSoar:
1. Start and AAT task in simulator mode
2. Cross the AAT task start cylinder
3. Select the MacCready Infobox
4. Hold the up button on the keyboard for several seconds
until the screen freezes.

Signed-off-by: Rob Dunning <rob@raspberryridgesheepfarm.com>