dlgTaskCalculator.cpp: do not hide Target button ever
allow use of the target button even for racing tasks.
This is consistent with previous versions, and allows the user
to quickly "scan" any task (racing or AAT) by zooming into
each of the turn points.
For Goto and Abort tasks, the task calculator and the
menu item for target are disabled, so this button need not
be hidden for these two types of tasks.