dltTarget.cpp: add verification that task is correct type
don't allow the dlgTarget to open for abort or goto tasks.
Even though we disable the menus that would lead to this
for those types of tasks, it's a good idea to double check
and not open the dialog because it does not show any info
for goto or abort tasks.