dlgTask{List,Point}: don't use Chart::DrawNoData()
This is no chart.  Drawing a white window is good enough, and if we
really want the "No data" text here, we should move the rendering code
out of the Chart class into a generic class.

Problem with Chart::DrawNoData() is that it assumes that the colors
and the font have been configured already.