dlgConfiguration: remove pages configuration
This is pretty useless currently, because this does not do anything
yet.  Anyway, users will not understand how this is intended, so
instead of trying to fix it before XCSoar 6.0, I'm removing it, and we
will revisit it for the next version.

At the same time, I'm removing Pages::PageLayout::MakeTitle() to
remove the translatable strings which we will never use anyway.
The profile loading code is only disabled by an #if.