DeviceBlackboard: Call flying_state_moving() when above 300m AGL
This fixes ticket #456

If we have valid terrain and are above 300m AGL without necessary speed we can assume to be moving in strong headwind. Previously the on_ground parameter was only taken into account for the calculation of the OnGround member, now the Flying member won't be reset anymore just because the speed isn't high enough e.g. in wave conditions.