dlgConfiguration: add an option to disable external wind
Previously, you had no chance to manually enter the wind when an
external device supplied us with a wind vector.  This new flag allows
you to disable the external wind source.

We could make that a separate option in AutoWindMode, but then we
would lose auto-switching when external wind becomes unavailable.

This implements ticket #426.