OLCFAI: optimisations (see notes)
No point scanning for start and finish, they can be assumed to be
closed (or eventually closed).  This reduces by an order of magnitude
the search size.

No point scanning triangle rules if total distance is less than previous
best.  This also reduces the number of expensive calls to distance
calculations.

Small consequence of these changes is we are not checking 1000m start-finish
elevation rule, nor will we have accurate start/finish stats.  This work
to be done later.