OLCFAI: fixed bugs (see comments)
Search performed in mixed order.  First point selected is the
most recent point in the flight, then following two points up to that
point are searched for.  This makes Dijkstra function correctly for
this special type of optimisation problem.

The solution order ends up wrong, but this won't effect much, and
can be fixed later.

To do still is accurate timing and closure test.