Revert "MapWindowRender: Removed #if 0 code"
This reverts commit 570238681b9a1da82749748531258be6f1126ea8.

This feature of having a backup artificial horizon based on inferred
orientation from GPS and vario data is useful, and reasonably well tested,
but has the issue of potentially invalidating use of XCSoar in FAI contests
due to rule ref Annex A to Section 3 (2010 Edition) 4.1.2 "No instruments
permitting pilots to fly without visual reference to the ground may be carried
on board, even if made unserviceable."  The quality of XCSoar's pseudo-AH
is arguably good enough that this violates the rule.  We need to seek clarification as to whether this is the case or not.