AATPoint/RenderTaskPoint: implement scoring deadzone
The samples polygon is not the best visual representation for AAT tps
of the area in which flight does not grant extra points for scoring;
this "scoring deadzone" is actually the convex hull formed from the
sampled polygon with the inclusion of the destination in the next
turnpoint (since you will be necessarily flying from somewhere inside
the sampled hull through to the destination).

This patch implements calculation of the deadzone on sample updates,
and adjusts the RenderTaskPoint class to draw the deadzone for current
points (previous points just use the sampled convex hull).  The name
of the functions *draw_samples are changed to *draw_deadzone to reflect
the new extended functionality.

Note that this patch makes the need, to either fix ticket #476 or
temporarily disable draw_deadzone, even more important.