XML: Fixed wrong value in GetPosition()
This fixes ticket #481

no_scaling was never set to true before. no_scaling must be set to true if the dialog scaling (fixed with, scaled, full, ...) should be skipped. This is essential if we use negative values. Simple example:

DlgWidth = 100
X= -4
X is set to 96
if X is now scaled with the dialog scale function it might become greater then 100 and moves completely out of sight.