SearchPoint: add constructor without projection
In the AirspacePolygon constructor, a dummy TaskProjection is passed,
because we don't know the projection yet at this point.  By not
passing the TaskProjection (with dummy values), we leave the projected
coordinates undefined, which is ok because they are not used or useful
at this point, and we don't need to calculate them.  Speed up by 15%.