Screen/Canvas: blit Bitmaps without BitmapCanvas
Every Canvas object has a "compatible_dc" variable which is created on
demand.  It is used to blit bitmaps internally, and the caller doesn't
need to use BitmapCanvas.

On SDL/OpenGL, BitmapCanvas wasn't needed anyway, so this is a big
simplification.

This patch is needed to give OpenGL better texture support.