AirspaceWarning: use WarningTime and AltWarningMargin
AirspaceWarningManager: make all predictions use config.WarningTime

Only update_task was using config.WarningTime, but all of the uses
of update_predicted should use it to be consistent.

AltWarningMargin moved to AirspaceWarningConfig

It's needed by the AirspaceWarningManager so the arbitrary 1000m
ceiling above the aircraft for predictions can be expanded if the
AltWarningMargin is greater than this value.