AbstractTask/OrderedTask: prevent stats jumping in OZ
For OZ's with non scored boundaries, the effective mc and cruise efficiency
stats suffer jumping from a discontinuity with scored distance.  This gives distracting information to the pilot.  This mechanism blocks updating of those statistics while in the OZ for non boundary OZs (AST turnpoints).

Tested with real flights for AST, works well.