dlgLoggerReplay: Switch ScanDirectoryTop() call order
We generate IGC files ourselves, so at some point the user may have very many of those files and at some point the currently static array is full so the NMEA files aren't added to the file list anymore. This is still not the perfect solution but it will give priority to the NMEA files.