Trace: new implementation
Old version was poorly implemented, too many lookups and time-based optimisation of tree was expensive.  This one is smart in that it uses sort on insert, so removing points is quick.

See documentation in Trace.hpp.  Scales very well now with size,
some scope for further minor optimisation but working well.

- use swap to avoid slow vector copy
- use mutable last_time in TracePoint so trace tree does not need to
  erase neighbour points to update last_time