AirspaceAltitude: rationalisation, AGL support
AbstractAirspace was directly accessing the altitude field of AirspaceAltitude.
This prevents use of the AirspaceAltitude struct to handle items such as AGL altitude.
So now moving accessors to AirspaceAltitude and make them private.

Use aircraft state as proxy for query state, allowing use of aircraft
state to determine the local terrain height for updating the AGL height reference for queries.
This makes an assumption/simplification that the entire airspace with AGL reference
is constant altitude with the terrain height being the terrain height of the aircraft.

Also changed here is use of AirspaceAltitude (baro preferred) to allow for finer control
over the altitude reference used by airspace queries.  Previously NavAltitude was used by some parts, Baro by others.  Now the altitude reference is defined by AirspaceAltitude, and is consistent.

AircraftStateFilter: filter AirspaceAltitude
- Need this for airspace medium/long term predictions.
- Add predicted Speed, Vario, remove TODO
- Remove unused m_alt