SampledTaskPoint: remove reference to TaskProjection
The TaskPoint classes should not carry a reference to the parent task
or its TaskProjection instance, because they don't need it most of the
time.  All modifications originate from the Task, and the caller can
pass a TaskProjection instance whenver the TaskPoint needs it.  This
removes overhead and makes the code a little bit safer, because
TaskPoint objects may now be portable.