Logger/IGCWriter: fall back to GPS altitude if baro unavailable
When there is no barometric altitude provided by the NMEA device, the
pressure altitude is retained at zero.  That confuses broken software
like SeeYou.

The IGC flight recorder specification does not mention what to do when
pressure altitude is not available.  It might be the best compromise
to just use the GPS altitude.  At least this is how XCSoar 5 worked.