GlueMapWindow: add flag to track minimum distance pan drag
This is used in PAN mode, if the Pan drag ever goes more than
about 10 pixels, then we disable mouse up events for waypoint
details and for airspace details.   This way if user clicks, drags
away and back to same point and release, it does not trigger the
details popup.