OLCSprint: dont update trace when futile
Since this is an online algorithm, all solutions must have start to end of trace
satisfy the finish altitude requirements.  otherwise there is no point
even retrieving the full trace or starting a search.

Assuming a bounded ceiling and very long flight, this would be expected to reduce
the number of trace acquisitions and solution starts by 50%.  In practice the number
will be lower than this but the fewer wasted cpu cycles the better.