OrderedTask: bypass expensive operations if far
We know during OrderedTask::check_transitions whether the aircraft
is near or far from the OZ, and can use this to bypass expensive
operations.  To do this, we split update_sample() into update_sample_near()
and update_sample_far().  If near, it is no guarantee that the aircraft
is inside, but if far, we know it is outside.

The operations that will be bypassed are polygon interior checks for AAT,
and for other OZ types, distance and sometimes bearing calculations.

The transitions calculations as well as update_samples calls the isInSector()
repeatedly, so bypassing these with foreknowledge that all are outside is useful.