Engine: implement next wp solution
Many users are confused by the fact that previously, XCSoar's
current leg solution, computed assuming you are continuing the task,
gives results for wp arrival distance etc that look strange --- the
users expect those figures to be as if you will land there.
We should not interfere with the engine's internals to adjust this,
but can just separately compute a solution to the active tp as if
the glider were to land there.

This patch implements those changes, and updates the infoboxes to
point to the new data.  Note that the solution is still computed at the
current mc, not safety mc.

Users have asked for the absolute arrival altitude of the waypoint, not
relative to safety height etc.

This provides an additional wp arrival altitude infobox.

This should fix tickets #458 and #578, #527