TaskManager: option to use safety polar for next_solution
Discussion on igc/trac has concerned whether the next wp data should
conform to the task (in which the current mc is used) or the safety polar
(in which the data will match what is shown on alt diff on the map).

This adds support (commented out) to switch to using the safety polar.