Dijkstra: allow overriding of default reserved queue size
This is required so when we know there is a temporary use of a Dijkstra,
we wont be allocating the whole storage each iteration.  This is used
in TaskDijkstra which creates a dijkstra object temporarily.
Ideally the TaskDijkstra object would be stored in the OrderedTask so could
benefit from the reserve allocation, but currently the TaskDijkstra
is initialised from ordered task information, so improving this work is to do.