GrahamScan: Reduce copying of vectors, minor fix
Reduce the GrahamScan object is temporary, there is no need
to pass a vector out of it, it can modify the input vector
directly if required.

Consequently, the use of the convex pruning to detect convexity
in SearchPointVector has to be updated to use a copy.  For future work,
the is_convex() shouldnt do any pruning at all but just do the scan part
of the GrahamScan.

Secondly, the use of size() everywhere is slow, it only needs to be performed once.
Thirdly, the ::equal test is not required, we will only copy the source vector
if the size was reduced.  If size wasnt reduced, it cannot have been non-convex.