GrahamScan/SearchPointVector: add tolerance to distance checks
This tolerance is required due to rounding errors causing sign() to fail for
points within a line, which is required for GrahamScan.  Providing for tolerance
allows this check to allow inclusion of slightly-convex points to be pruned, thereby
simplifying the polygon.

Also added method thin_to_size, which may be used later to ensure
the SearchPointVector size (and therefore min/max distance calcs) is bounded.