TaskPoint: add ROUTE taskpoint type
This will be used for path planning nodes.  Nothing uses it yet,
but this provides for UI routines to know how to identify such
items for rendering in a different way to normal TaskWayPoints.

Visitors are not yet set up to handle ROUTE types, which will
probably have their own task point type eventually.