Screen/BufferWindow: repaint in on_paint() each time
Don't repaint internally once and paint it many times; this is not how
the Window is used and how it is useful.  This class is documented
only as "avoids flickering", its goal is not to reduce the number of
on_paint_buffer() calls.  Remove the method repaint(), callers should
use invalidate() instead.

This might fix #652, because the drawing is not done by the
InstrumentThread anymore.