Airspace: support for route planner
m_border was used only by AirspacePolygon, but we want circles to have a
border so we can use get_search_points() for route planning.

Airspaces: method to copy within range

This copy does not own the airspaces, it is just a subset of the
airspaces within range.  The actual airspaces themselves are not copied,
only the containers.

This will be used for path planning.

Airspaces: allow caller access to projection

Airspace: each airspace has a clearance polygon

Management of this is on demand by get_clearance() calls,
to clear, call clear_clearance() in the individual or clear_clearances()
in the Airspaces database.

Airspaces: synchronise methods

These are used to obtain a subset database that can be persistent
but periodically synchronised to add/delete items if the query results
change.

This is used by the AirspaceRoute planner.

Airspace: == operator allows for direct tree erasing
Otherwise find() method scans through the tree
Airspaces: erase_exact is faster

AirspacePolygon: use positive/negative where possible

Safer, we can do eps checking if required later.

AirspaceCircle: use 12 sided polygon for border representation