ThermalBase: more stable algorithm
Iterating over time is unstable due to the dt being non-monotonic in some conditions,
in particular when terrain altitude may increase as we trace back to the source.
This would happen if you leave a thermal just downwind of the top of a hill.

A more accurate and stable way of tracing back is to descend in altitude, so
height variations do not adjust step size until the intercept is found.