GlideComputerTask: fix mc setting loops
Disable setting risk mc to devices temporarily (see comments in code).
Ideally we want to send the risk mc to devices but not have it set as the true
global mc setting as this produces a decay feedback loop.

This addresses #583.

Also check for MC changed by the computer, just because we are in auto mode,
the computer may not adjust the mc value, so allow the computer to take mc
from the blackboard.

This may address #498.