OrderedTask: improved start vector logic
The system previously locked out the smart positioning of the virtual start point
(that produces shortest distance to start rather than directing the
pilot through the start location), if the task had already been started.

There is no need to be so conservative, the engine properly remembers the
true start location, so the smart start vector can be retained in all modes.
Secondly, this patch is more careful about resetting the start on an invalid exit.