Airspaces: reset QNH on insert so FLs are updated
QNH was initialised to zero only on class construction, and is set on set_flight_levels
to prevent over-active set_flight_levels on members.  However, this meant that
after the database was cleared, the flight levels would not be reset (the QNH trigger
would fail).  So now, on insert we force m_QNH to be set to zero.