xmlParser, DataNode: use TextWriter and TextLineReader
Serialise the XML into the file, without allocating a temporary
buffer.  This changes the character encoding from ACP (what is that,
anyway?) to UTF-8.  TextLineReader is used to guarantee that the
character set is well-defined while reading.  UTF-8 is needed because
ACP is very limited, and cannot support many locales.

Without the temporary string, there is no allocation and no memory
leak.  Fixes TRAC ticket #687.