RTTaskFactory.cpp: remove unique turnpoint constraint for RT
Remove the unique turnpoint constraint from racing t asks.  This would
prevent a pilot from using XCSoar in a contest where a legal racing task
required repeating a turnpoint.

The unique turnpoint requirement applies to FAI distance tasks
only.  It does not apply to RT tasks.
FAI Sporting code section 3 appendix A says:
6.3.1 Racing Task: The Organisers shall set a Start, two or
more Turn Points (7.5.1) to be achieved in order, and a Finish.

The only mention of using a turnpoint only once is in SC3 sections
1.4.5 and 1.4.7b which refers to distance tasks for badges, diplomas
and records.